[FIX]Refactor this module and fix the readme/pep8/header.
