+++ hr_timesheet_line_readonly/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# © 2015 Elico corp (www.elico-corp.com)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
