+++ hr_timesheet_line_readonly/__openerp__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+# © 2015 Elico corp (www.elico-corp.com)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+{
+    'name': 'Hr Timesheet Line Readonly',
+    'version': '8.0.1.0.1',
+    'category': 'hr',
+    'depends': [
+        'hr_timesheet',
+    ],
+    'author': 'Elico Corp',
+    'license': 'AGPL-3',
+    'website': 'https://www.elico-corp.com',
+    'data': [
+        'hr_timesheet_line_readonly_view.xml'
+    ],
+    'installable': True,
+    'application': False,
+}
