+++ project_issue_id_display/__openerp__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# © 2015 Elico corp (www.elico-corp.com)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    'name': 'Project Issue ID Display',
+    'version': '8.0.1.0.1',
+    'category': 'Project',
+    'depends': [
+        'project_issue'
+    ],
+    'author': 'Elico Corp',
+    'license': 'AGPL-3',
+    'website': 'https://www.elico-corp.com',
+    'data': [
+        'project_issue_id_display.xml'
+    ],
+    'installable': True,
+    'application': False,
+}
