+++ project_planning/project_task_view.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+    <data>
+        <record id="view_task_form2" model="ir.ui.view">
+            <field name="name">project.task.form</field>
+            <field name="model">project.task</field>
+            <field name="inherit_id" ref="project.view_task_form2"/>
+            <field name="arch" type="xml">
+                <field name="progress" position="after">
+                    <field name="planned_hours" widget="float_time" groups="project.group_time_work_estimation_tasks" on_change="onchange_planned(planned_hours, effective_hours)"/>
+                    <field name="priority" groups="base.group_user" widget="priority"/>
+                </field>
+                <notebook position="before">
+                    <table>
+                        <tr>
+                            <th><label for="week_1"/></th>
+                            <th><label for="week_2"/></th>
+                            <th><label for="week_3"/></th>
+                            <th><label for="week_4"/></th>
+                        </tr>
+                        <tr>
+                            <td><field name="week_1" nolabel="1"/></td>
+                            <td><field name="week_2" nolabel="1"/></td>
+                            <td><field name="week_3" nolabel="1"/></td>
+                            <td><field name="week_4" nolabel="1"/></td>
+                        </tr>
+                    </table>
+                </notebook>
+            </field>
+        </record>
+        <record id="view_task_tree2" model="ir.ui.view">
+            <field name="name">project.task.tree</field>
+            <field name="model">project.task</field>
+            <field name="inherit_id" ref="project.view_task_tree2"/>
+            <field name="arch" type="xml">
+                <field name="remaining_hours" position="after">
+                    <field name="week_1" string="P1"/>
+                    <field name="week_2" string="P2"/>
+                    <field name="week_3" string="P3"/>
+                    <field name="week_4" string="P4"/>
+                </field>
+            </field>
+        </record>
+    </data>
+</openerp>
