+++ project_task_id_display/project_task_view.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+    <data>
+        <record id="view_task_kanban" model="ir.ui.view">
+            <field name="name">project.task.view.task.kanban</field>
+            <field name="model">project.task</field>
+            <field name="inherit_id" ref="project.view_task_kanban"/>
+            <field name="arch" type="xml">
+                <field name="name" position="before">
+                        <field name="id"/>
+                </field>
+            </field>
+        </record>
+    </data>
+</openerp>
