+++ project_task_project_mandatory/__openerp__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# © 2015 Elico corp (www.elico-corp.com)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    'name': 'Elico Project',
+    'version': '8.0.1.0.1',
+    'category': 'Project',
+    'depends': [
+        'project_timesheet',
+        'project_issue_sheet'
+    ],
+    'author': 'Elico Corp',
+    'license': 'AGPL-3',
+    'website': 'https://www.elico-corp.com',
+    'installable': True,
+    'application': False,
+}
