+++ project_task_type/__openerp__.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+# © 2015 Elico corp (www.elico-corp.com)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+{
+    'name': 'Project Task Type',
+    'version': '0.0.1',
+    'category': '',
+    'depends': [
+        'project_timesheet',
+        'project_issue_sheet'
+    ],
+    'author': 'Elico Corp',
+    'license': 'AGPL-3',
+    'website': 'https://www.elico-corp.com',
+    'demo': ['project_demo.xml'],
+    'data': [
+        'project_task_view.xml',
+    ],
+    'test': [],
+    'installable': True,
+    'application': False
+}
