+++ project_task_work_bottom_edit/__openerp__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# © 2015 Elico corp (www.elico-corp.com)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    'name': 'Task Work Bottom Edit',
+    'version': '8.0.1.0.1',
+    'category': 'Project',
+    'depends': [
+        'project_timesheet',
+    ],
+    'author': 'Elico Corp',
+    'license': 'AGPL-3',
+    'website': 'https://www.elico-corp.com',
+    'data': [
+        'project_task_view.xml',
+    ],
+    'installable': True,
+    'application': False
+}
