+++ project_task_work_bottom_edit/project_task_view.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+    <data>
+        <record id="view_task_form2" model="ir.ui.view">
+            <field name="name">project.task.form</field>
+            <field name="model">project.task</field>
+            <field name="inherit_id" ref="project.view_task_form2"/>
+            <field name="arch" type="xml">
+                <xpath expr="//tree" position="attributes">
+                      <attribute name="editable">bottom</attribute>  
+                </xpath>         
+            </field>
+        </record>
+    </data>
+</openerp>
